data: fixing data type of min and max temp
I forgot to check the definition area of the min and max temperature of the data file. So I fixed it now. I replace the double data type to int.

+++ src/main/java/de/exxcellent/challenge/data/WeatherData.java
@@ -19,10 +19,10 @@ public class WeatherData {
     /**
      * Property of the maximum of temperature of the day.
      */
-    private double maxTemp;
+    private int maxTemp;
 
     /**
      * Property of the minimum of temperature of the day.
      */
-    private double minTemp;
+    private int minTemp;
 }
